+++ app/src/main/java/me/portailler/florian/testanimation/ui/tabbar/state/TabBarItem.kt
@@ -1,0 +1,9 @@
+package me.portailler.florian.testanimation.ui.tabbar.state
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class TabBarItem(
+	val icon: ImageVector,
+	val selectedIcon: ImageVector = icon,
+	val label: String?,
+)